# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/credit-applications/confirmation": {
            "post": {
                "description": "Подтверждение заявки на кредит",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credit"
                ],
                "summary": "Подтверждение заявки на кредит",
                "parameters": [
                    {
                        "description": "CreditApplicationConfirmationV1Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreditApplicationConfirmationV1Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Заявка на кредит подтверждена",
                        "schema": {
                            "$ref": "#/definitions/model.CreditApplicationConfirmationV1Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/credit-applications/{id}": {
            "get": {
                "description": "Получение заявки на кредит",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credit"
                ],
                "summary": "Получение заявки на кредит",
                "responses": {
                    "200": {
                        "description": "Заявка на кредит получена",
                        "schema": {
                            "$ref": "#/definitions/model.GetCreditApplicationV1Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Обновление статуса заявки на кредит",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credit"
                ],
                "summary": "Обновление статуса заявки на кредит",
                "parameters": [
                    {
                        "description": "UpdateCreditApplicationStatusV1Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateCreditApplicationStatusV1Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Статус заявки на кредит обновлен",
                        "schema": {
                            "$ref": "#/definitions/model.UpdateCreditApplicationStatusV1Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/credits/{id}": {
            "get": {
                "description": "Получение информации о кредите",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credit"
                ],
                "summary": "Получение информации о кредите",
                "responses": {
                    "200": {
                        "description": "Информация о кредите получена",
                        "schema": {
                            "$ref": "#/definitions/model.GetCreditV1Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/credits/{id}/payment-schedule": {
            "get": {
                "description": "Получение расписания платежей",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Получение расписания платежей",
                "responses": {
                    "200": {
                        "description": "Расписание платежей получено",
                        "schema": {
                            "$ref": "#/definitions/model.GetPaymentScheduleV1Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CreditApplicationConfirmationV1Request": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
        "model.CreditApplicationConfirmationV1Response": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.CreditV1": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "credit_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "interest_rate": {
                    "type": "integer"
                },
                "remaining_amount": {
                    "type": "integer"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "model.GetCreditApplicationV1Response": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "approved_amount": {
                    "type": "integer"
                },
                "decision_date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.GetCreditV1Response": {
            "type": "object",
            "properties": {
                "credit": {
                    "$ref": "#/definitions/model.CreditV1"
                }
            }
        },
...
        "model.GetPaymentScheduleV1Response": {
            "type": "object",
            "properties": {
                "payments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PaymentSchedule"
                    }
                }
            }
        },
...
        "model.PaymentSchedule": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateCreditApplicationStatusV1Request": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "decision_notes": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.UpdateCreditApplicationStatusV1Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}